Infer Groovy 3 jars from classpath if found, otherwise declare external dependencies
Groovy 3 is not repackaged in groovy-all jar as was the case for previous versions.
When inferring Groovy3 classpath for tools (like groovyc), we will never have
the -all artifact, thus we have to look for its individual pieces
(possibly added via `localGroovy()`) and use those if found.
Otherwise, declare external dependencies (as was the case for Groovy <3
when no -all artifact was found on classpath).