Use hierarchical accessors for versions and bundles
For consistency with the `libraries` accessors, versions and bundles
now use the same structure, which is to say that a kebab-case alias
is converted into nested accessors.

Previous to this commit, a version with an alias `foo-bar-baz` was
converted into the accessor `libs.fooBarBaz`, but a library with
the same alias would have been converted to `libs.foo.bar.baz`.

Now both behave the same, with nested accessors.

Fixes #16676