Test latest and earliest versions by default in multi-version tests
Currently, if no specific version to test is given for a multi-version test,
for example, when running embeddedIntegTest without -PtestVersions argument,
only one version is picked up. It is supposed to be the latest one.
This change proposes an alternative - always test first and last supported versions
in the list to get the feedback in the quick test pipeline (or locally).