Always provide a clear recompile reason
The optimization of a dependencyToAll with no reason was not necessary.
There are only very few classes that fall into that category, namely
module-info and source retention annotations. Storing a reason for those
leads to much better usability.

We also track where changed constants come from now, so that the users knows
exactly which class triggered the full recompile. This comes at virtually no
cost, since this state is only in memory for the duration of the compilation
and we abort class analysis on the first full recompile reason.

As part of this, I've overhauled the tests for constant changes to only describe
the current behavior, not some potential future behavior. This removes a lot of
tests that were just asserting that something doesn't work. If we fix constant
tracking, we'll definitely know, this won't happen by accident.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>