Remove custom settings files from samples
Those were meant to be used with now deprecated command line flag.
This will also make the sample clearer as the usage of this custom
settings file is not mentioned anywhere in the sample documentation.
Those were only really needed for tests, yet they appear in the
downloadable sample, which can confuse users.
The functionality itself should be tested outside of the sample.