Remove kotlin-dsl buildscript compilation avoidance test workaround
This workaround relied on debug output to assert whether the buildscripts or plugins blocks
were recompiled or not. This was because BuildOperationsFixture was failing an internal
assertion when precompiled script plugins were used at the time of writing.
Whatever was the root cause of failures in BuildOperationsFixture, it is now gone
and the workaround can be removed.