Ignore Kotlin daemon deprecation warning on JDK16 on all executors
On JDK16, Gradle sets a deprecated JVM argument --illegal-access=permit for
the Kotlin compiler daemon (this will be fixed in Kotlin 1.5 line).
Some of our tests rely on --info output where this deprecation message
can be logged when a new daemon JVM happens to be spawned, making them flaky.
Initial assumption was that this happens only in no-daemon mode, but
some recent flakiness examples show that this is not the case.
Thus, for now, do not do deprecation checks on JDK16 in Kotlin DSL tests
that rely on --info output.