Replace AtomicReference by volatile field
The original intent was to avoid duplicating/missing
events when the MutableFileWatchingStatistics gets
swapped out in getAndResetStatistics() at the end of
the build. However, it is still possible to record
events duplicated like this.

This PR replaces the AtomicReference by a volatile field.
That simplifies things. It is still possible to miss events,
though that is not so important, given that these are
statistics anyway.

See #16794