Use `IntSet` instead of `Set<Integer>`
This commit makes sure to use `IntSet` during incremental Java compilation
instead of `Set<Integer>`. IntSets are already used in parts of the code
which are executed in the daemon, but it wasn't the case for  code executed
in a compiler daemon. To make this possible the fastutil library is now
exported to workers which require the Java compiler.