Display plugin in validation messages
This commit adds more context to type validation problems, by
trying to identify to which plugin a type belongs. Unfortunately
there's no reliable way to associate a particular type to a plugin.

The way this change works is that whenever a type is validated, we
identify the jar to which it belongs. Then we look at the loaded
plugins, and for each known plugin class, we identify the jar they
are defined in. If we find that the type belongs to the same jar,
then we ask Gradle in which plugin the plugin class is defined and
use this as the plugin id.

This wouldn't work if, for example, a plugin uses multiple jars and
that the type we're validating is not in the _same jar_ as the plugin
itself.

If a jar contains multiple plugins, then we would pick the first one
which matches.

Last but not least, if a jar actually only contains types and no
concrete `Plugin<...>` implementation (for example if a plugin only
contribute task types), then we wouldn't be able to identify the
owner of a type.

Fixes #16904