Check that the Java home for the potential compatible daemon still exists
When a Gradle build runs with JDK-A, we insert this into the daemon registry as a
potential compatible daemon for later.

If JDK-A is then deleted but the registry entry still exists, subsequent Gradle
builds will try to see if they are compatible with this entry and fail because
they cannot find JDK-A anymore.

This change checks that the JDK installation directory still exists before probing
further.

fixes #16816