Shrink incremental compilation state files
Deduplicate class and resource names across whole PreviousCompilationData
-------------------------------------------------------------------------

Before this change, we had duplicate class names in the classpath snapshot,
output snapshot, annotation processing data and constants mapping. This was
bloating the previous compilation data file significantly.

With this change, each name is only written once and reused across all
these data structures.

Merge sourceClassesMapping into previousCompilationData
-------------------------------------------------------

This removes an inefficient plain text file and replaces it with
an efficient binary representation. It also avoids duplicate storage
since the class names are already contained in the previous compilation
data, so the only thing we're newly adding are the relative source paths.
These are now optimized with our hierarchical name serializer.

As part of this change, the Groovy and Java compiler have been aligned to
both use the ApiCompilerResult, which opens up the path for incremental
annotation processing and constant tracking in Groovy.

There is an inconsistency in how the two languages handle class files
for which the source could not be determined. Groovy does a full recompile,
Java just ignores it. It turns out that the Java case is because we don't
determine the relative path for files generated by annotation processors.
I've fixed half of the issue by adding the generated sources dir to the
CompilationSourceDirs. However, there is also the possibility to generate
class files directly from an annotation processor and in that case our class
name collector will not see that class. In order to keep this scenario incremental,
I've accepted the inconsistent behavior for now. This should be fixed in a later
change, so that annotation processing correctly tracks generated class files
and both Java and Groovy do a full recompile for any files that are untracked.

The CompilerApiData now distinguishes between whether it supports only source to
class mapping or whether it also supports constant mapping. This is required to
make Groovy work correctly, as Groovy does not yet collect constant data, but does
collect source to class data.

Simplify ConstantDependentsMapping
----------------------------------

Tracking the visited classes is unnecessary. It was previously used in the
merging code, but this code really only needed the set of changed classes,
which was already given by the RecompilationSpec

Replace IntSet with DependentSet, as we already have efficient serialization
for those and the Strings on the compiler API side are deduplicated already.

Make hierarchical name serialization denser
-------------------------------------------

Split names along package separators, nested class separators, file separators
and camel case borders. Give each prefix/suffix an integer ID and write that ID
whenever the same prefix/suffix is encountered in a name.

Use constants analysis when minifying classpath snapshots
---------------------------------------------------------

Now that we have proper constant tracking, we no longer need to keep all classes
on the classpath which have constants, but only those that were actually used.

Correct handling of array types in ClassDependenciesVisitor
-----------------------------------------------------------

I noticed during profiling that we were sometimes referencing array types in our
class analyis, which we should never do. I traced this back to the inconsistent
handling in the ClassDependenciesVisitor. I've simplified this so that handling
of array and primitive types is now centralized, so that they are always unwrapped
or ignored correctly.

Results
========

- gradle/gradle now uses 20MB, down from 49.1 MB. Could be even less if we implemented
constant tracking for Groovy, as that would allow us to discard all unused constant
holders form the analysis file.
- largeMonolithicJavaProject now uses 10.6 MB, down from 23 MB

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>