Use the module name by default for artifact selectors
Previously to this commit, the `artifact { ... }` notation in a dependency
required the user to set the `name`. In fact, we can use the dependency
name by default: it's likely to be the case. So we're basically turning
what used to be an error (build wouldn't pass) into something which is
expected.

Fixes #16768