Support ignoring bucketed performance test
Previously, we can't simply ignore the performance test because that
would make bucket JSON inconsistent with the performance tests, which
fails sanityCheck. But sometimes we do want to ignore a performance
test temporarily - that's why this PR exists.

This PR introduces a `@IgnorePerformanceTest` annotation, which adds
a `ignore` field to JSON. The ignored performance test will still be put
into bucket, but at the runtime it will be skipped.