Use a test-only lock for testing for lock contention
Instead of relying on a particular lock we know Gradle will
ask for, creating a lock that is only used by this test so
we have more control over when the lock is requested and
released.

This should not change any behavior in the test, just the
lock that is used in it.