Add Gradle build smoke test that would fail on deprecations
This traverses only a small path in our build logic catching any deprecations
only in that path.
To really catch all possible deprecations in our own build logic we should
run nearly the whole pipeline using newly built Gradle distribution and exercise
as much of the build logic as possible. This might not be feasible.