Introduce an abstraction for services that report on problems encountered during the build execution.
For now, the only thing this abstraction supports is to allow problem reporting to be wired into the correct stage of the build action pipeline in a common way, rather than the ad hoc wiring currently used for each different thing that reports problems. It also allows some common infrastructure to be extracted and shared for all problem reporting in the future and edges cases to be fixed in a single place.

Use this to report configuration cache problems that are generated during tooling model problems.