Rework test that explicitly needs TLSv1 or TLSv1.1 support
Java removed support for TLSv1 and TLSv1.1 in JDK16 and this was
backported to all other versions.

https://bugs.java.com/bugdatabase/view_bug.do?bug_id=JDK-8202343

To make this test work again in the same way, we would need to either
re-enable TLSv1/1.1 support in the _test worker_, which requires running
the test worker with special flags or modifying the JDK's
java.security file. Or the test would need to run the HTTPS server in
another process (with TLSv1/1.1 support).

To test the same effect (changing https.protocols changes the
protocols used by the build), we only enable TLSv1.2 on the server and
explicitly ask for TLSv1.3. This will fail the build.