Enhance the Java version metadata available for toolchains
The `java.version` property might not be unique enough to identify
a toolchain. It is replaced with `java.runtime.version` which includes
a build number (it is useful for EA releases). The `java.vm.version` is
also made available so it is possible to report usages of non-Hotspot
VMs like J9 or GraalVM.

Fixes #17085