Use installation path to break ties when selecting java toolchain
Some vendors sometimes ship JDKs that only differ in vendor-specific
version properties (like java.vendor.version) or without differences at
all. If such JDKs are installed alongside then prior to this patch the
selection was nondeterministic and different JDKs might be selected for
different runs causing inconsistencies.

With this patch an installation path (which is guaranteed to be unique)
of the JDK is used to break ties. The selected JDK is still as arbitrary
as its installation path but it is always the same for the same machine
(until another JDK is installed, of course).