Disallow included plugin builds to reach out to root build of the product
This prevents build logic of a plugin build from accessing and potentially
modifying the build logic of the root build (the build that's building
the product), which should be done by applying the plugins that the
included build produces instead.

The problem of included plugin builds accessing the root build became
evident when the plugin build is searched for settings plugins and
thus evaluated before the root build is ready. This made it possible
for an unrelated and transitively applied plugin (that tries to reach
the root of the whole build tree) somewhere in the
build logic of the plugin build fail the main build if the build
is searched for settings plugins.