Do not set a context class loader when creating thread.
When creating a thread, the current context class loader (if any) will
be used to set the initial context class loader for the thread.

In Gradle's case, the initial class loader is the VisitableClassLoader
instance for the current build.

Because thread usages will always set/reset the context class loader
before/after any work actions, the initial context class loader
will be kept forever during the lifetime of the daemon.

Issue #16797

Signed-off-by: Jerome Dochez <jedo@google.com>