Update test for verification of javadocs/sources only
Dependency verification is trigerred _when the file is accessed_. That is
to say that while an artifact resolution query may trigger download of
dependencies and files, those files would only be _verified_ if, and only if,
the build actually accesses those files.

Without calling `it.get(SourcesArtifact)*.file`, we wouldn't trigger
dependency verification.