Add an option to disable key servers
This commit provides a usability improvement to dependency verification. Normally, Gradle
would first look at the local keyring file for keys, and if a key isn't found, it would
start searching for the key in remote servers. However, key servers tend to be flaky and
may trigger build failures on CI if a key isn't in the local keyring.

In order to make sure that the local keyring always contains the whole set of keys that
is used in a build, there's now an option to disable lookup in key servers. For this, the
XML file has to be configured with `<key-servers enabled="false"/>`.

This is **not** the same as setting an empty key server list, since an empty list would
mean that we use the default list of key servers instead.

In addition, this fixes a problem with the dependency verification XML schema, which
lacked the `key-servers` section!