Remove restriction on naming of aliases
In previous releases, it wasn't possible to declare a version catalog alias which was an
intermediate node, but also provided a value. For example, this was invalid (using TOML
but the API had the same restrictions):

   [libraries]
   top="org:foo:1.0"
   top.bottom="org:bar:1.0"

The reason is that we need the leaves to be providers and there's no custom implementation
of a provider for dependency notations (custom implementations are expensive to maintain
and should be avoided).

To workaround this problem, this commit introduces an interface, `ProviderConvertible`,
which is implemented by intermediate nodes which also provide a value. This interface
provides a single method, `asProvider()`, which is transparently called by Gradle when
such a notation is added as a dependency:

    implementation(libs.top) // libs.top implements ProviderConvertible

so this is equivalent to writing:

    implementation(libs.top.asProvider())

And accessing the leaf is possible as before using:

    implementation(libs.top.bottom)

Fixes #17170