Fix "Multiple builders are available" when syncing project in IDEA
There is a workaround that prevents adding a builder multiple times:
https://git.io/JZLn1
This workaround relies on the fact that there is the `builders` field in
the BuilderRegistry. The field was replaced with `registrations` in
a recent commit so the workaround no longer works.

This commit introduces a quick fix by providing `builders` field that
is backed by the data in `registration`.

Proper fix will probably require figuring out why the plugin is applied
twice to the same Gradle instance.