Add support for version catalogs in `plugins { ... }` block
This commit adds the ability to declare plugins in version catalogs. In a similar
way to libraries, plugins are declared with an alias, but point to a plugin id
instead of (group, artifact).

The notation to use a plugin declared in a catalog is similar to the one used
to declare dependencies:

```
plugins {
   id(libs.plugins.somePlugin)
}
```

it is possible to override the version declared in a catalog:

```
plugins {
   id(libs.plugins.somePlugin) version "1.5"
}
```

This works for both the Groovy and Kotlin DSLs.