Add support for plugins in TOML file
In addition to the settings API, this commit adds support for plugins in TOML
files.

While currently plugins only support a single (required) version, the infrastructure
is ready for using richer version constraints, so the TOML file (and the settings API)
support rich version constrainsts. Only the required version would be used though.