Replace task execution listener by simple tasks for code quality build scan data extraction
For compatibility with the configuration cache

This commits also make the reported paths relative to the root directory
instead of relative to each project directory.

This is now using allprojects {} and should later be replaced by
applying a plugin to all code-quality projects instead to comply with
isolated projects constraints.

Signed-off-by: Paul Merlin <paul@gradle.com>