Remove TransformationResult.resultsInWorkspace
Since we now only store the absolute paths to
produced outputs in the TransformationResult,
we don't need to check if the input artifact
is as well in the global cache. Wherever it is,
we will restore the correct location relative to
it from the cache. And the produced outputs for
the ImmutableWorkspaceProvider will always be in
the global cache, so we don't need to use an extra
check.