Reduce number of Gradle version*JDK version combinations
Gradle 2.6 has bugs that causes flakiness in this test. We're mostly
interested in knowing that the current version of the TAPI with various
combinations of Gradle, daemon JDK versions and TAPI JDK versions works
as expected.

We have some coverage for Java 6 builds working with a non-Java 6 TAPI.

We don't have to test the minimum version of Gradle that works with
everything when a less buggy version is available.