Remove commented-out sample for jacoco aggregation of all Test tasks
The commented out code that is shown in the sample is not correct. tasks.withType(Test) will not realize all the test tasks, but rather iterate over an empty list and yield nothing in this case if the tasks have not yet been configured. An alternative would be to wrap this code in afterEvaluate {}.