Move the state that is tracked for tasks that are a dependency of tasks in other builds in the tree so that the state is managed by the `BuildState` instance that owns the task.
This allows the state to be queried directly from the task node instance instead of being duplicated outside the `BuildState` instance. This also moves the state closer to being tracked by the task graph for the build (which manages the nodes). This also removes some further differences between the root and included builds.

This commit also moves the "task graph ready" build operations of each included build to run only after all tasks have been discovered. These operations move out of the "calculate task graph" build operation for the child build to become children of the build operation of the root build.