Fix performance of Micronaut annotation processing
This reverts 5f6aa44665389f1e703bb24d788ad5ee305404db and fixes the actual root cause of the problem.
The issue was that we were not deleting aggregated types/resources on file additions, because our addition
logic ignored the "types/resources depending on all others". This problem was introduced with SourceFileClassNameConverter,
since we no longer infered the name of newly added types and instead completely ignored the transitive dependency analysis
on additions.

This change adjusts our aggregating processor fixture to only generate a resource file, just like Micronaut does. The reason
for this is that there simply is no way to write a valid aggregating processor that creates a Java type. This is because new
annotated types can appear each processing round, so the aggregating processor has to wait until `processingOver()` is true
to catch all of them. But if the registry class is generated after `processingOver`, then it can't be referenced anywhere
and the compiler issues a warning that a Java class was generated in the final round. Generating a resource does not have
this issue, since resources aren't referenced by Java types anyway, they are just loaded at runtime.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>