Simplify/Fix ClassSetAnalysisData
The recursive search now terminates as soon as a full recompile reason is found and
returns this reason. It also handles transitive dependencies correctly again, which
was broken at some point during the introduction of `privateDependentClasses`. Whenever
a class was both privately and publicly referenced, the public reference was ignored,
leading to missed recompilations.

Fixing these issues allowed me to remove a bunch of special-case code, especially around
constant dependencies, greatly simplifying the algorithm.

Another simplified are are annotation processing dependencies. We no longer recompile the
source file when the generated file is affected by something. Instead, we only reprocess the
source file so that the generated file is recreated. This is more correct and efficient and
removes the boilerplate code around `dependenciesFromAnnotationProcessing`.

I've added additional test coverage and documentation to keep this from regressing.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>