Don't start another daemon process from launcher when the first one is not accessible
Partial fix for #12232

In some network configurations, it can happen that the daemon client can start a new daemon but the new process cannot be connected. The client, however, does not distinguish between new and existing daemon connections and performs retries. Since the retry limit is set to 100, the client spams the entire memory with inaccessible Gradle daemons.

The solution is to make the client know if a new daemon was started. If the connection fails in that case then the client does not do further retries. This is also the behavior of the single-use daemons.

This PR also documents the use-case and steps for the investigation.