Add full compile classpath of a sourceset to PMD's auxclasspath
The compile classpath of the sourceset doesn't entirely consist of the
compileClasspath configuration. It may also include other things; for
example the compile classpath of the test source set contains outputs
of the compilation of the main source set. These things are crucial
for the PMD analysis too.

This CL adds all non-configuration-originated entries of the compile
classpath to the PMD aux classpath.