ExecutionPlan: acquire necessary locks before attemting to collect mutation info
When the collection of task outputs on a node triggers a configuration resolution,
we want to make sure that the lock on the project is acquired before.

If not, the DefaultConfiguration.resolveToStateOrLater method will attempt to
get a lock, and in so doing will free all other ExecutorWorker waiting to
find a new node. This leads to executionQueue iterators from different threads
to work at the same time with possible "ConcurrentModificationException" occurring.

Signed-off-by: Emmanuel Guerin <e.guerin@criteo.com>