Add support for Scala 3
Issue: #16527

Previously, we would need to compile the Scala compiler bridge in order to use it with a particular Scala version. Now, it's possible to download the jar, which is released together with a particular Scala version.

From other changes, a new compiler is needed together with Scala library, however it still reuses the Scala 2.13 library underneath.

Sample project this works with:

```kotlin
plugins {
    scala
}

repositories {
    mavenCentral()
    mavenLocal()
}

dependencies {
    implementation("org.scala-lang:scala3-library_3:3.0.1")
}
```

Signed-off-by: Tomasz Godzik <tomek.godzik@gmail.com>