Add a resource download service
This commit adds a resource download service which can be used by
toolchain provisioning to list, for example, available JDKs, then
perform the download itself.

The remote resource service honors the traditional offline mode
and performs caching of resources. It's worth noting that toolchain
provisioning calls the provisioning methods twice, but this isn't
a new behavior: it's just exercised by this pull request, which adds
tests with fake HTTP servers, highlighting this behavior.

Signed-off-by: Cedric Champeau <cedric.champeau@oracle.com>