Fix performance at scale of IterationOrderRetainingSetElementSource#add
Fixes #16998

This `add` method was taking time linear in the size of the collection
rather than constant time, causing performance issues at scale.

There is a slight semantic change here: Previously, if an element was
submitted using `add`, it could be rejected if it matched an element
from a provider, but only if the provider had been realized. This
means the outcome of the iterator could have changed based on the
timing of this realization, and was also different from what the
behavior would be if the element were added through a provider.

Now, if provider and non-provider versions of the same value are
added, they will generally both be accepted into the collection
at `add` time (with a return value of true), and at iteration
time, only one of them will be returned.

Signed-off-by: Alex Landau <alandau@palantir.com>