Make cross-version JDK TAPI test work when Java 6 is not installed
Instead of failing with a NPE for all tests when Java 6 is not
installed, the test can use Java 8 and compile to Java 6 bytecode.

This means we only need to skip execution when Java 6 is not
installed for Gradle 2.14.1 and we can run all the other tests
with only Java 8+ installed.