Remove restriction on subgroups in version catalogs for plugins
This removes restriction on subgroups in version catalogs for plugins. Fix for libraries was already applied before but for plugins this did not work.

So now if you define plugin aliases: greeter and greeter-second, this can be normally used in plugins block as:
   plugins {
      alias(libs.plugins.greeter)
      alias(libs.plugins.greeter.second)
   }

To make this work ProviderConvertible<T> had to be moved to public API.

This commit also adds integration tests that similar functionality works for buildscripts.

Fixes #18142