Invalidate the VFS for untracked outputs
We did hide untracked inputs and outputs completely from the execution engine.
We actually don't have to, since the problematic parts is the snapshotting of
both outputs and inputs. And we don't do this any more as soon as we detect any
untracked input or output.

Whether we pass the untracked input or not doesn't make any difference - we
don't do anything apart from fingerprinting with the input anyway. For the
outputs, the story is different. Especially the step where we invalidate the VFS
for an output location before the task start executing is something we also
should do for untracked inputs.

Passing the untracked outputs to the execution engine causes the following:
- We create the output directories.
- We invalidate the VFS for the outputs before starting task execution.