Fix BuildOperationTrace with debug on Java 8
The first log event which ends up in BuildOperationTrace initializes the Groovy
VM on the call of JsonOutput.toJson(). That is not bad as is, only on Java 8
this causes the logging of an exception while initializing the VMPlugin. And
from the debug logging, we end up in the trace again which tries to call
JsonOutput.toJson again. Since the VMplugin isn't initialized, yet (we are just
in the middle of it), we end up with an NPE and the build fails.

So we need to initialize the VMPlugin before we start writing traces.

If the log level is not debug, the Groovy VM doesn't log the error, so we don't
come into this situation.