Introduce resolution strategy to deactivate global substitution rules
Gradle implicitly registers dependency substitution rules for all
configurations in the whole build tree to find projects in other
included builds. These rules are always active by default.

There are however cases, where a certain configuration should not apply
these rules when resolving. For example, if a binary version of a module
should be discovered that is also represented by a project in the same
or another build.

One use case for this is to compare the previous published Jar of
a module with the latest locally built Jar of the same module.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>