Acquire locks before evaluating mutation info
When the collection of task outputs on a node triggers configuration
resolution, we want to make sure that the lock on the project is acquired before
 so there are no `ConcurrentModificationException`s
or deadlocks.

Note that a better way to do this would be to resolve the
mutations in a separate node and move the resolution out
of the coordination lock.

#17905
Replaces #17906