Change the configuration cache entry layout to separate the file system watch directories and the fingerprint into different files.
Write the file system watch directories at the end of the build, rather than when a value is written to the cache entry. There can potentially be multiple values written to the cache entry, for example when both tasks and tooling model creation are run, or when tooling models are cached per project.