Introduce ArtifactCollection.getResolvedArtifacts()
It returns a Provider to allow access to the list of resolved artifacts in a lazy manner.
The provider is backed by the actual ArtifactColletion.
It is working with the configuration cache out of the box thanks to the existing ArtifactCollection support.
Later changes will allow to use this as a task input.

Signed-off-by: Paul Merlin <paul@gradle.com>