Replace enumeration of the environment variables with explicit list
The filtering approach required enumerating all available environment
variables. This makes all these variables inputs to the configuration
cache so any change will invalidate the cache once the tracking code for
environment variables.

With this commit the enumeration of the variables and name-based
filtering is replaced with the explicit list of variables because of the
limitations of the provider API. Currently, it isn't possible to ask for
a subset of environment variables. Unfortunately, this makes adding new
environment variables a bit harder. However, the new API to request a
subset of variables based on the prefix is currently considered so we
can move to it when it is available.