Add the variant identity to the seenArtifacts key
In ResolvedArtifactCollectingVisitor, the artifacts are identified by
ResolvableArtifact.getId(), so in case of
ComponentFileArtifactIdentifier the artifact is identified by the
componentId and fileName which could be the same for artifacts in the
same module (the same ProjectComponentIdentifier) and producing the
same fileName (but the path is different).

This change adds the variant attributes and capabilites as part of the
key in ResolvedArtifactCollectingVisitor to solve this case.

Signed-off-by: Amr Afifiy <amrmahmoud@google.com>