Allow the build scan plugin to read environment variables
Some versions of the build scan plugin read environment variables during
configuration phase but do not take decisions upon them, so recording
these variables as inputs to the configuration cache is false positive.
Similar to system properties, we skip these access. The escape hatch to
disable the workarounds is provided to make fixing the issues in the
scan plugin easier in the future. Workarounds for environment variables
and system properties can be turned off separately.