Add special handling for aggregating and enumerating methods
Some methods of maps returned by System.getProperties() and
System.getenv() do not work with concrete keys. Instead, these methods
aggregate the whole contents of the map (like size() or hashCode()) or
expose the contents to the caller (like iterator() or forEach()). If
such a method is called during the configuration phase then any
change in the environment (environment variable or system property)
can change the behavior of the caller. The configuration cache has no
idea what changes affect the configuration and what don't. Therefore,
the only way to keep correctness is to add a fingerprint of the whole
environment to the cache, and we should discourage build authors from
using such patterns to improve cache hit rates.

This commit moves us closer to the described ideal by fingerprinting
every property/variable if an aggregating or enumerating method is
called. This approach is quicker to implement than adding a full-blown
special fingerprint. While not all environment changes are detected
(e.g. adding a variable will not invalidate the cache), some are and
the sheer number of inputs can highlight to the build author that
they are doing something fishy.