Don't reprocess derived types if primary type is reprocessed
If AHelper is an annotated class that was generated from another annotated class A,
and both need to be reprocessed because of a change, then only A should be passed
to the compiler for processing. AHelper will be recreated as part of the processing
of A. Passing AHelper explicitly results in javac throwing a duplicate type error.

Fixes #18840

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>