Make Kotlin editor support work when exception happen in afterEvaluate block
The Kotlin DSL IntelliJ integration needs to evaluate the build script to provide
context for editing .gradle.kts files. This feature needs to work even if the build
script results in runtime exceptions.

This PR introduces the ExceptionCollector build service that suppresses
and collects exceptions that happen in afterEvaluate callbacks. The suppressing
implementation is loaded if Gradle is invoked with the following JVM flag:

org.gradle.kotlin.dsl.provider.mode=classpath

If that's not the case, a no-op service implementation is loaded and the Gradle
configuration works as before.

For more context see https://github.com/gradle/gradle/issues/18637