Fix SnapshotTaskInputsBuildOperationResult.visitInputFileProperties
for ignored root directories.

`visitUnvisitedDirectories` didn't reset the state, so the
root directory was reported twice instead of reporting the
subdirectory. This is only a problem now since we changed
relative path normalization to always ignore the root directory,
so we have a situation where only some directories are ignored
and not all of them.