Let Gradle create a "mirror" directory with downloaded artifacts
This commit introduces a system property, `org.gradle.download.mirror.directory`,
which, when set, would make Gradle copy all resources which it downloads during
the build into a local directory, in addition to its internal cache.

The rationale is to make it easier for external tooling to create mirrors of
repositories which are used during a build. Indeed, unlike Maven, Gradle copies
downloaded resources into an opaque cache directory structure (for good reasons),
but making it difficult to create a repository from scratch which would include
all the required dependencies.

The goal is slightly different from the 2-level dependency cache for CI: in this
case the goal is NOT to create an environment cache, but to create a trusted
repository that an organization can verify independently, track every single
dependency or file being downloaded and mirror into, say, an Artifactory or
Nexus server, which would then be used on _different_ builds.

There were several attempts in doing so. For example:

- https://github.com/uklance/gradle-dependency-export which cannot handle plugins
and has to deal with parent POMs/BOMs in a special manner since those are not
accessible via standard Gradle APIs
- https://github.com/melix/gradle-download-helper which does something similar
using the tooling API but has to download files by itself in a separate context,
while having the same problem with parent POMs and BOMs

This commit provides a _simple_ solution to this problem, which, while not perfect,
makes it much easier to create such a layout:

- you would run a build with a fresh Gradle home (-g newhome)
- and pass the `org.gradle.download.mirror.directory` property to a local directory

As soon as a file is downloaded, it would automatically be copied into that directory.
At the end of the build, you get a "local mirror" which could potentially be used
as a file repository or used as a source for uploading to a mirror.

Signed-off-by: Cedric Champeau <cedric.champeau@oracle.com>