Allow applying plugins with matching versions
This resolves a long-standing issue with applying a plugin in a
subproject when that same plugin is already applied in the root project.
Prior to this commit, you would need to avoid specifying a version,
which only became a true hassle with the introduction of plugins in the
version catalog, as shown in #18236.

This commit does not attempt to resolve conflicting versions, or inspect
the classpath for a plugin version. It only records the data that Gradle
knows for sure, and uses that to slightly relax this check. It is
sufficient to allow the issue mentioned prior to be resolved.