Let system properties use providers as values
This commit adds support for using providers as values of system
properties for `JavaForkOptions`. This basically covers the main
`JavaExec`, `Test` and `ExecOperations.javaexec` use cases.

In practice, before, any system property which needed to be
computed lazily had to be set in a `doFirst` block, which is far
from being perfect. In addition, tracking the right inputs was
difficult.

With this change, it's now possible to use providers as values
and the inputs of the task are properly tracked. This basically
offers the ability to do things like:

```
tasks.named('test', Test) {
    systemProperty("foo", tasks.named('bar').map { it.outputFile.asFile().get().text })
}
```

Fixes #12247

Signed-off-by: Cedric Champeau <cedric.champeau@oracle.com>